1120 fix spectrum scale for no data (#1455)
* Explode MarkerLegend/index into individual files

* Fix look of empty legend

* Fix no data scale

* temp

* Revert temp and fix single value legend

* Organise imports and clean up

* Reuse code

* Organise imports

* Rename mapStateToProps and use cleaner syntax


* Wrap components in React.memo

* Clean up rendering empty legend

* Move propTypes above mapStateToProps

* Redo NoDataLabel <- functional change

* Rename file to LegendContainer.js

* Remove comment

* change normaliseToPercentage logic

* Fix regression
