+++ packages/web-frontend/src/components/Marker/markerColors.js
@@ -65,10 +65,12 @@ export function resolveSpectrumColour(scaleType, scaleColorScheme, value, min, m
 }
 
 const normaliseToPercentage = (value, min = 0, max = 1) => {
-  const normalisedValue = (value - min) / (max - min);
-
   // Always clamp the result between 0 and 1
-  return Math.max(0, Math.min(1, normalisedValue));
+  if (value < min) return 0;
+  if (value > max) return 1;
+  if (min === max) return 0.5;
+
+  return (value - min) / (max - min);
 };
 /**
  * Takes a value and return a hsl color string for use as a style
