+++ packages/web-frontend/src/containers/MeasureLegend/LegendContainer.js
@@ -1,0 +1,14 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ */
+
+import styled from 'styled-components';
+
+export const LegendContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: center;
+`;
